perf: do not flush maps on COMM for perf report
Patchwork-id: 58089
O-Subject: [PATCH RHEL6.5 BZ927866 260/264] perf tools: do not flush maps on COMM for perf report
Bugzilla: 927866
RH-Acked-by: Don Zickus <dzickus@redhat.com>
RH-Acked-by: Ivan Vecera <ivecera@redhat.com>
RH-Acked-by: Prarit Bhargava <prarit@redhat.com>

From: Luigi Semenzato <semenzato@chromium.org>

Bugzilla: 927866
https://bugzilla.redhat.com/show_bug.cgi?id=927866

upstream status
===============
commit 9fdbf671ba7e8adb2cbb93d716232ebcab55f6bd
Author: Luigi Semenzato <semenzato@chromium.org>
Date:   Tue Aug 21 14:52:20 2012 -0700

description
===========
This fixes a long-standing bug caused by the lack of separate COMM and EXEC
record types, which makes "perf report" lose track of symbols when a process
renames itself.

With this fix (suggested by Stephane Eranian), a COMM (rename) no longer
flushes the maps, which is the correct behavior.  An EXEC also no longer
flushes the maps, but this doesn't matter because as new mappings are created
(for the executable and the libraries) the old mappings are automatically
removed.  This is not by accident: the functionality is necessary because DLLs
can be explicitly loaded at any time with dlopen(), possibly on top of existing
text, so "perf report" handles correctly the clobbering of new mappings on top
of old ones.

An alternative patch (which I proposed earlier) would be to introduce a
separate PERF_RECORD_EXEC type, but it is a much larger change (about 300
lines) and is not necessary.