ipmi: refer to table as "SPMI", not "ACPI"
Patchwork-id: 58164
O-Subject: [RHEL6.5 PATCH 01/21] ipmi: refer to table as "SPMI", not "ACPI"
Bugzilla: 924736
RH-Acked-by: Jes Sorensen <Jes.Sorensen@redhat.com>
RH-Acked-by: Myron Stowe <mstowe@redhat.com>
RH-Acked-by: Ivan Vecera <ivecera@redhat.com>
RH-Acked-by: Neil Horman <nhorman@redhat.com>

Cherry-picked from the following upstream commit.

commit 18a3e0bfbcd589599d0affbfd484ba9a97e5f122
From: Bjorn Helgaas <bjorn.helgaas@hp.com>
Date: Tue, 17 Nov 2009 17:05:29 -0700

This discovery method uses the SPMI table, not the ACPI namespace.  In
the future, we will look in the namespace, so let's refer to the table
as "SPMI" and save "ACPI" for the namespace.

Signed-off-by: Bjorn Helgaas <bjorn.helgaas@hp.com>
Signed-off-by: Len Brown <len.brown@intel.com>
Signed-off-by: Jarod Wilson <jarod@redhat.com>