ipmi: use locks on watchdog timeout set on reboot
Patchwork-id: 58183
O-Subject: [RHEL6.5 PATCH 20/21] ipmi: use locks on watchdog timeout set on reboot
Bugzilla: 924736
RH-Acked-by: Jes Sorensen <Jes.Sorensen@redhat.com>
RH-Acked-by: Myron Stowe <mstowe@redhat.com>
RH-Acked-by: Ivan Vecera <ivecera@redhat.com>
RH-Acked-by: Neil Horman <nhorman@redhat.com>

Cherry-picked without conflicts from the following upstream commit.

commit 423a5bb49ec530ec8bbfc73fd2ded83da8e58684
Author: Corey Minyard <cminyard@mvista.com>
Date:   Wed Mar 28 14:42:50 2012 -0700

ipmi: use locks on watchdog timeout set on reboot

The IPMI watchdog timer clears or extends the timer on reboot/shutdown.
It was using the non-locking routine for setting the watchdog timer, but
this was causing race conditions.  Instead, use the locking version to
avoid the races.  It seems to work fine.

Signed-off-by: Corey Minyard <cminyard@mvista.com>
Signed-off-by: Andrew Morton <akpm@linux-foundation.org>
Signed-off-by: Linus Torvalds <torvalds@linux-foundation.org>
Signed-off-by: Jarod Wilson <jarod@redhat.com>