async: Make async execution api a little more fair in scheduling
Patchwork-id: 58190
O-Subject: [RHEL6.5 PATCH bz871955] async: Make async execution api a little more fair in scheduling
Bugzilla: 871955
RH-Acked-by: Don Zickus <dzickus@redhat.com>

Bug 871955 reported that we had a problem with some third party developers
debugging storage systems.  Specifically, if they enabled scsi debugging and did
something involving heavy load, like forcing a rescan of several hundred luns,
we'd wind up getting lots of "CPU stuck" warnings and eventually locking up the
box for various and sundry reasons.  Its not a huge concern since we don't enable
scsi debugging in production system, but it still seemed odd especially since
they reported the same problem didn't occur when they used an upstream kernel.

All the lockups had a common data point in that they were all prefaced with
heavy logging that originated from paths leading from the async api (functions
scheduled with the async_schedule call).  As it turns out, upstream has modified
this bit of infrastructure to use the system_unbound workqueue,  Which has lots
of code in place to balance its execution with the needs of other tasks and
execution contexts in any given running system.

This wouldn't be a big deal to backport, were it not for the fact that pulling
back the changes to implement system_unbound_wq amount to completely updating
the workqueue api, which we can't do as it would cause major abi breakage.

So instead I've done this, it makes four changes to the current async api
implementation that should reduce the overall amount of work it places on the
system when lots of load is applied to it, and keep us from starving other
tasks:

1) Lowers MAX_THREADS.  Currently MAX_THREADS is hardcoded to 256, which makes
very little sense.  On systems with lots of cpus, theres no point in not using
them, while on systems with less than 256 cpus, we just create a huge amount of
thread contention.  Instead we should set MAX_THREADS to the number of online
cpus minus 1, leaving at least one cpu available for other work, but using at
least one (in the up case).

2) Fix a thread exit counting bug.  There is a race in the thread retirement
path, where a thread marks itself as exiting, but can return if it detects work
gets added to the queue first.  This can lead to the async api creating more
tasks than MAX_THREADS would otherwise allow.  Change that to use
atomic_add_unless to create proper co-ordination between threads in async_thread
that are exiting with the creation of new threads in async_manager_thread.

3) Fix a task state bug.  We constantly set our task state to TASK_INTERRUPTIBLE
in async_thread, even if we don't sleep. Only do that if we're actually going to
schedule a timeout for a period.

4) Introduce a forced schedule when the kernel flags us to do so.  Functions
that we run may (a) run for a long time, and (b) may do so with pre-emption
disabled, leading to us executing past our scheduling quanta (use of spinlocks
bh and irq variants will use preempt_enable_no_resched).  When this happens we
can monopolize the cpu.  So, after executing each work item, a thread can check
need_resched, and appropriately call schedule then, to keep us fair.

BZ: 871955
Brew: http://brewweb.devel.redhat.com/brew/taskinfo?taskID=5190254
Testing: Tested successfully by the reporter
Signed-off-by: Jarod Wilson <jarod@redhat.com>