xfs: pass shutdown method into xfs_trans_ail_delete_bulk
Patchwork-id: 58204
O-Subject: [RHEL6 PATCH] xfs: pass shutdown method into xfs_trans_ail_delete_bulk
Bugzilla: 805407
RH-Acked-by: Brian Foster <bfoster@redhat.com>
RH-Acked-by: Dave Chinner <dchinner@redhat.com>

Bugzilla:
https://bugzilla.redhat.com/show_bug.cgi?id=805407

upstream:
04913fdd91f342e537005ef1233f98068b925a7f

Brew:
https://brewweb.devel.redhat.com/taskinfo?taskID=5599646

xfs_trans_ail_delete_bulk() can be called from different contexts so
if the item is not in the AIL we need different shutdown for each
context.  Pass in the shutdown method needed so the correct action
can be taken.

In contrast with upstream there are a few different places from where
xfs_trans_ail_delete() is called:

Callers in upstream that are not in rhel6:
xfs_qm_dqflush()

Callers in rhel6 that are not in upstream
xfs_inode_free

    Signed-off-by: Dave Chinner <dchinner@redhat.com>
    Reviewed-by: Christoph Hellwig <hch@lst.de>
    Reviewed-by: Mark Tinguely <tinguely@sgi.com>
    Signed-off-by: Ben Myers <bpm@sgi.com>

Signed-off-by: Carlos Maiolino <cmaiolino@redhat.com>
Signed-off-by: Jarod Wilson <jarod@redhat.com>