be2iscsi: cleanup a min_t() call
Patchwork-id: 58213
O-Subject: [RHEL6.5 PATCH 01/37] [scsi] be2iscsi: cleanup a min_t() call
Bugzilla: 947550
RH-Acked-by: Tomas Henzl <thenzl@redhat.com>

From: Dan Carpenter <dan.carpenter@oracle.com>

Bugzilla: 947550

"sense_len" was declared as int type but actually it only stores a
u16 value that comes from hardware.  The cast to u16 in min_t()
confuses static analysis because it truncates the int to u16 so I've
fixed the declaration to reflect that "sense_len" is just a u16.

Also there was a call to cpu_to_be16() which I've changed to
be16_to_cpu().  The functions are equivalent, but obviously the
hardware is big endian and we're doing the min_t() comparison on CPU
endian values.

This whole patch is just a cleanup and doesn't affect how the code
works.

http://marc.info/?l=linux-scsi&m=132260052803847&w=2

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>
Reviewed-by: Mike Christie <michaelc@cs.wisc.edu>
Acked-by: Jayamohan Kallickal <Jayamohan.kallickal@emulex.com>
Signed-off-by: James Bottomley <JBottomley@Parallels.com>
Signed-off-by: Jarod Wilson <jarod@redhat.com>