pSeries_lpar_hpte_remove fails from Adjunct partition being performed before the ANDCOND test
Patchwork-id: 58366
O-Subject: [PATCH RHEL6.5 BZ952309 v2] powerpc: pSeries_lpar_hpte_remove fails from Adjunct partition being performed before the ANDCOND test
Bugzilla: 952309
RH-Acked-by: Ivan Vecera <ivecera@redhat.com>
RH-Acked-by: Jes Sorensen <Jes.Sorensen@redhat.com>
RH-Acked-by: Stefan Assmann <sassmann@redhat.com>

RHBZ#:
------
https://bugzilla.redhat.com/show_bug.cgi?id=952309

Description:
------------
Some versions of pHyp will perform the adjunct partition test before the
ANDCOND test.  The result of this is that H_RESOURCE can be returned and
cause the BUG_ON condition to occur. The HPTE is not removed.  So add a
check for H_RESOURCE, it is ok if this HPTE is not removed as
pSeries_lpar_hpte_remove is looking for an HPTE to remove and not a
specific HPTE to remove.  So it is ok to just move on to the next slot
and try again.

Cc: stable@vger.kernel.org
Signed-off-by: Michael Wolf <mjw@linux.vnet.ibm.com>
Signed-off-by: Stephen Rothwell <sfr@canb.auug.org.au>

v2 - Fix spaces instead of tabs issue that Stefan pointed out.

RHEL Version Found:
-------------------
RHEL 6.4

Brew:
-----
https://brewweb.devel.redhat.com/taskinfo?taskID=5673589

Upstream:
---------
http://git.kernel.org/cgit/linux/kernel/git/torvalds/linux.git/commit/?id=9fb2640159f9d4f5a2a9d60e490482d4cbecafdb

Test Status:
------------
Mike Wolf helped me test this patch.

---------------------------------------------------------------
Steve Best
IBM on-site partner

Proposed Patch:
---------------

Signed-off-by: Jarod Wilson <jarod@redhat.com>