gfs2: Add "top dir" flag support
Patchwork-id: 58445
O-Subject: [RHEL6.5 PATCH 1/9] GFS2: Add "top dir" flag support
Bugzilla: 859446
RH-Acked-by: Abhijith Das <adas@redhat.com>
RH-Acked-by: Steven Whitehouse <swhiteho@redhat.com>

This patch adds support for the "top dir" flag. Currently this is unused
but a subsequent patch is planned which will add support for the
Orlov allocation policy when allocating subdirectories in a parent
with this flag set.

In order to ensure backward compatible behaviour, mkfs.gfs2 does
not currently tag the root directory with this flag, it must always be
set manually.

rhbz#859446

Signed-off-by: Jarod Wilson <jarod@redhat.com>