gfs2: dont reference inode:s glock during block allocation trace
Patchwork-id: 58452
O-Subject: [RHEL6.5 PATCH 8/9] GFS2: don't reference inode's glock during block allocation trace
Bugzilla: 859446
RH-Acked-by: Abhijith Das <adas@redhat.com>
RH-Acked-by: Steven Whitehouse <swhiteho@redhat.com>

This patch changes the block allocation trace so that it references
the rgd's glock rather than the inode's glock. Now that the order
of inode creation is switched, this prevents a reference to the
glock which may not be set yet.

rhbz#859446

Signed-off-by: Jarod Wilson <jarod@redhat.com>