gfs2: Flush work queue before clearing glock hash tables
Patchwork-id: 58462
O-Subject: [RHEL6.5 PATCH][bz#952836] GFS2: Flush work queue before clearing glock hash tables
Bugzilla: 952836
RH-Acked-by: Abhijith Das <adas@redhat.com>
RH-Acked-by: Steven Whitehouse <swhiteho@redhat.com>

Hi,

This patch is for bz952836 - kernel panic gfs2_glock_put from glock_workqueue process

There was a timing window when a GFS2 file system was unmounted
that caused GFS2 to call BUG() and panic the kernel. The call
to BUG() is meant to ensure that the glock reference count,
gl_ref, never gets down to zero and bounce back up again. What was
happening during umount is that function gfs2_put_super was dequeing
its glocks for well-known files. In particular, we saw it on the
journal glock, sd_jinode_gh. The dequeue caused delayed work to be
queued for the glock state machine, to transition the lock to an
"unlocked" state. While the work was still queued, gfs2_put_super
called gfs2_gl_hash_clear to clear out the glock hash tables.
If the timing was just so, the glock work function would drop the
reference count at the time when it was being checked for zero,
and that caused BUG() to be called. This patch calls
flush_workqueue before clearing the glock hash tables, thereby
ensuring that the delayed work is executed before the hash tables
are cleared, and therefore the reference count never goes to zero
until the glock is cleared.

rhbz#952836

The patch was tested on system gfs-a16c-01.mpc.lab.eng.bos.redhat.com
with an additional patch that opened the timing window wide enough
to recreate the problem. Without this patch, the problem would recreate.
With the patch, it would not.

The patch has been accepted for the nmw upstream git tree, but
hasn't made its way there yet.

Regards,

Bob Peterson
Red Hat File Systems

Signed-off-by: Bob Peterson <rpeterso@redhat.com>
Signed-off-by: Jarod Wilson <jarod@redhat.com>