cifs: turn read_from_socket into a wrapper around a vectorized version
Patchwork-id: 58501
O-Subject: [RHEL6 PATCH 20/55] [bz734728]cifs: turn read_from_socket into a wrapper around a vectorized version
Bugzilla: 734728
RH-Acked-by: Jeff Layton <jlayton@redhat.com>

Eventually we'll want to allow cifsd to read data directly into the
pagecache. In order to do that we'll need a routine that can take a
kvec array and pass that directly to kernel_recvmsg.

Unfortunately though, the kernel's recvmsg routines modify the kvec
array that gets passed in, so we need to use a copy of the kvec array
and refresh that copy on each pass through the loop.

Reviewed-and-Tested-by: Pavel Shilovsky <piastry@etersoft.ru>
Signed-off-by: Jeff Layton <jlayton@redhat.com>
(cherry picked from commit 42c4dfc213190fafffc53815c2ee6064430bc379)

Signed-off-by: Sachin Prabhu <sprabhu@redhat.com>
Signed-off-by: Jarod Wilson <jarod@redhat.com>