cifs: add a callback function to receive the rest of the frame
Patchwork-id: 58500
O-Subject: [RHEL6 PATCH 28/55] [bz734728]cifs: add a callback function to receive the rest of the frame
Bugzilla: 734728
RH-Acked-by: Jeff Layton <jlayton@redhat.com>

In order to handle larger SMBs for readpages and other calls, we want
to be able to read into a preallocated set of buffers. Rather than
changing all of the existing code to preallocate buffers however, we
instead add a receive callback function to the MID.

cifsd will call this function once the mid_q_entry has been identified
in order to receive the rest of the SMB. If the mid can't be identified
or the receive pointer is unset, then the standard 3rd phase receive
function will be called.

Reviewed-and-Tested-by: Pavel Shilovsky <piastry@etersoft.ru>
Signed-off-by: Jeff Layton <jlayton@redhat.com>
(cherry picked from commit 44d22d846fdc7c3e688fc1ff5ae6d06d08bb5656)

Signed-off-by: Sachin Prabhu <sprabhu@redhat.com>
Signed-off-by: Jarod Wilson <jarod@redhat.com>