cifs: ensure that we always do cifsFileInfo_get under the spinlock
Patchwork-id: 58531
O-Subject: [RHEL6 PATCH 55/55] [bz734728] cifs: ensure that we always do cifsFileInfo_get under the spinlock
Bugzilla: 734728
RH-Acked-by: Jeff Layton <jlayton@redhat.com>

The readpages bug is a regression that was introduced in 6993f74a5.
This also fixes a couple of similar bugs in the uncached read and write
codepaths.

Also, prevent this sort of thing in the future by having cifsFileInfo_get
take the spinlock itself, and adding a _locked variant for use in places
that are already holding the lock. The _put code has always done that
so this makes for a less confusing interface.

Cc: <stable@vger.kernel.org> # 3.5.x
Reviewed-by: Pavel Shilovsky <pshilovsky@samba.org>
Signed-off-by: Jeff Layton <jlayton@redhat.com>
Signed-off-by: Steve French <smfrench@gmail.com>
(cherry picked from commit 764a1b1acecedfe204cb2e80d8e2cc7c6df1b0b8)

The patch corresponding to the one mentioned in the summary which
introduced the problem in the rhel6 devel tree is
32672e0652f8f4dc2e02a9d027deb4edb167d1cc

Signed-off-by: Sachin Prabhu <sprabhu@redhat.com>
Signed-off-by: Jarod Wilson <jarod@redhat.com>