vfs: fix spurious return of ESTALE errors when cwd on nfs is renamed on server
Patchwork-id: 58641
O-Subject: [RHEL6.5 PATCH] [BZ#896580] vfs: fix spurious return of ESTALE errors when cwd on nfs is renamed on server
Bugzilla: 896580
RH-Acked-by: Alexander Viro <aviro@redhat.com>
RH-Acked-by: J. Bruce Fields <bfields@redhat.com>

This patch is a "loose" backport of the upstream patch that fixes the
same problem. This backport is quite different from the upstream version
due to differences in the path-walking code and so that we can preserve
kABI.

For RHEL6, we can't reasonably kill FS_REVAL_DOT altogether, and must
add a new FS_WEAK_REVALIDATE flag in order to indicate that the
filesystem provides this operation.

Upstream patch description that also describes the problem follows.

------------------------[snip]----------------------

vfs: kill FS_REVAL_DOT by adding a d_weak_revalidate dentry op

The following set of operations on a NFS client and server will cause
ESTALE to be returned inappropriately:

    server# mkdir a
    client# cd a
    server# mv a a.bak
    client# sleep 30  # (or whatever the dir attrcache timeout is)
    client# stat .
    stat: cannot stat `.': Stale NFS file handle

Obviously, we should not be getting an ESTALE error back there since the
inode still exists on the server. The problem is that the lookup code
will call d_revalidate on the dentry that "." refers to, because NFS has
FS_REVAL_DOT set.

nfs_lookup_revalidate will see that the parent directory has changed and
will try to reverify the dentry by redoing a LOOKUP. That of course
fails, so the lookup code returns ESTALE.

The problem here is that d_revalidate is really a bad fit for this case.
What we really want to know at this point is whether the inode is still
good or not, but we don't really care what name it goes by or whether
the dcache is still valid.

Add a new d_op->d_weak_revalidate operation and have complete_walk call
that instead of d_revalidate. The intent there is to allow for a
"weaker" d_revalidate that just checks to see whether the inode is still
good. This is also gives us an opportunity to kill off the FS_REVAL_DOT
special casing.

Cc: NeilBrown <neilb@suse.de>
Signed-off-by: Jeff Layton <jlayton@redhat.com>
Signed-off-by: Al Viro <viro@zeniv.linux.org.uk>
Upstream commit: ecf3d1f1aa74da0d632b651a2e05a911f60e92c0

Signed-off-by: Jarod Wilson <jarod@redhat.com>