sched: Avoid cputime scaling overflow
Patchwork-id: 58699
O-Subject: [RHEL6 PATCH 1/2] sched: Avoid cputime scaling overflow
Bugzilla: 912662
RH-Acked-by: Prarit Bhargava <prarit@redhat.com>
RH-Acked-by: Frederic Weisbecker <fweisbec@redhat.com>
RH-Acked-by: Larry Woodman <lwoodman@redhat.com>

BZ#912662

Description (mostly from Frederic Weisbecker):
Some users have reported that after running a process with
hundreds of threads on intensive CPU-bound loads, the cputime
of the group started to freeze after a few days.

This is due to how we scale the tick-based cputime against
the scheduler precise execution time value.

We add the values of all threads in the group and we multiply
that against the sum of the scheduler exec runtime of the whole
group.

This easily overflows after a few days/weeks of execution.

A proposed solution to solve this was to compute that multiplication
on stime instead of utime, RHEL6 commit:

ddeed9c8ecc41369a27d86cd4b19d6a477e50538
"[kernel] cputime: Avoid multiplication overflow on utime scaling"

The rationale behind that was that it's easy for a thread to
spend most of its time in userspace under intensive CPU-bound workload
but it's much harder to do CPU-bound intensive long run in the kernel.

This postulate got defeated when a user recently reported he was still
seeing cputime freezes after the above patch. The workload that
triggers this issue relates to intensive networking workloads where
most of the cputime is consumed in the kernel.

This patch avoids stime * rtime overflow by dropping precision when
the numbers are big. Since the stime:utime ratio is already an
approximation through the sampling, reducing its resolution isn't a
big problem.

Upstream:
commit 55eaa7c1f511af5fb6ef808b5328804f4d4e5243
Author: Stanislaw Gruszka <sgruszka@redhat.com>
Date:   Tue Apr 30 17:14:42 2013 +0200

    sched: Avoid cputime scaling overflow

    Here is patch, which adds Linus's cputime scaling algorithm to the
    kernel.

    This is a follow up (well, fix) to commit
    d9a3c9823a2e6a543eb7807fb3d15d8233817ec5 ("sched: Lower chances
    of cputime scaling overflow") which commit tried to avoid
    multiplication overflow, but did not guarantee that the overflow
    would not happen.

    Linus crated a different algorithm, which completely avoids the
    multiplication overflow by dropping precision when numbers are
    big.

    It was tested by me and it gives good relative error of
    scaled numbers. Testing method is described here:
    http://marc.info/?l=linux-kernel&m=136733059505406&w=2

Brew:
https://brewweb.devel.redhat.com/taskinfo?taskID=5712801

Testing:
Compiled corresponding code in user-space and test it against big
values of rtime/stime/total - overflow does not occur. Patched kernel
was tested by running various programs that utilize utime/stime values,
no regression found.

Signed-off-by: Jarod Wilson <jarod@redhat.com>