gfs2: Clean up gfs2_create
Patchwork-id: 58799
O-Subject: [RHEL6.5 PATCH 1/7] GFS2: Clean up gfs2_create
Bugzilla: 859444
RH-Acked-by: Abhijith Das <adas@redhat.com>

If we pass through knowledge of whether the creation is intended to be
exclusive or not, then we can deal with that in gfs2_create_inode
and remove one set of locking. Also this removes the loop in
gfs2_create and simplifies the code a bit.

rhbz#859444