gfs2: Remove bd_list_tr
Patchwork-id: 58798
O-Subject: [RHEL6.5 PATCH 2/7] GFS2: Remove bd_list_tr
Bugzilla: 859444
RH-Acked-by: Abhijith Das <adas@redhat.com>

From: Steven Whitehouse <swhiteho@redhat.com>

This is another clean up in the logging code. This per-transaction
list was largely unused. Its main function was to ensure that the
number of buffers in a transaction was correct, however that counter
was only used to check the number of buffers in the bd_list_tr, plus
an assert at the end of each transaction. With the assert now changed
to use the calculated buffer counts, we can remove both bd_list_tr and
its associated counter.

This should make the code easier to understand as well as shrinking
a couple of structures.

rhbz#859444