nfsd: clean up and clarify the cache expiration code
Patchwork-id: 58787
O-Subject: [RHEL6.5 PATCH 11/34] [BZ#681606] nfsd: clean up and clarify the cache expiration code
Bugzilla: 681606
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>
RH-Acked-by: J. Bruce Fields <bfields@redhat.com>

Add a preprocessor constant for the expiry time of cache entries, and
move the test for an expired entry into a function. Note that the current
code does not test for RC_INPROG. It just assumes that it won't take more
than 2 minutes to fill out an in-progress entry.

I'm not sure how valid that assumption is though, so let's just ensure
that we never consider an RC_INPROG entry to be expired.

Signed-off-by: Jeff Layton <jlayton@redhat.com>
Signed-off-by: J. Bruce Fields <bfields@redhat.com>
Upstream commit: d1a0774de6cb908f5ba7806d09aaf86bb03fa182

Signed-off-by: Jarod Wilson <jarod@redhat.com>