nfsd: dynamically allocate DRC entries
Patchwork-id: 58781
O-Subject: [RHEL6.5 PATCH 15/34] [BZ#681606] nfsd: dynamically allocate DRC entries
Bugzilla: 681606
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>
RH-Acked-by: J. Bruce Fields <bfields@redhat.com>

The existing code keeps a fixed-size cache of 1024 entries. This is much
too small for a busy server, and wastes memory on an idle one.  This
patch changes the code to dynamically allocate and free these cache
entries.

A cap on the number of entries is retained, but it's much larger than
the existing value and now scales with the amount of low memory in the
machine.

Signed-off-by: Jeff Layton <jlayton@redhat.com>
Signed-off-by: J. Bruce Fields <bfields@redhat.com>
Upstream commit: 0338dd157282c19696d3c32614a748d0ba814b12

Signed-off-by: Jarod Wilson <jarod@redhat.com>