sunrpc: trim off trailing checksum before returning decrypted or integrity authenticated buffer
Patchwork-id: 58790
O-Subject: [RHEL6.5 PATCH 22/34] [BZ#681606] sunrpc: trim off trailing checksum before returning decrypted or integrity authenticated buffer
Bugzilla: 681606
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>
RH-Acked-by: J. Bruce Fields <bfields@redhat.com>

When GSSAPI integrity signatures are in use, or when we're using GSSAPI
privacy with the v2 token format, there is a trailing checksum on the
xdr_buf that is returned.

It's checked during the authentication stage, and afterward nothing
cares about it. Ordinarily, it's not a problem since the XDR code
generally ignores it, but it will be when we try to compute a checksum
over the buffer to help prevent XID collisions in the duplicate reply
cache.

Fix the code to trim off the checksums after verifying them. Note that
in unwrap_integ_data, we must avoid trying to reverify the checksum if
the request was deferred since it will no longer be present when it's
revisited.

Signed-off-by: Jeff Layton <jlayton@redhat.com>
Upstream commit: 4c190e2f913f038c9c91ee63b59cd037260ba353

Signed-off-by: Jarod Wilson <jarod@redhat.com>