nfsd: fix startup order in nfsd_reply_cache_init
Patchwork-id: 58791
O-Subject: [RHEL6.5 PATCH 26/34] [BZ#681606] nfsd: fix startup order in nfsd_reply_cache_init
Bugzilla: 681606
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>
RH-Acked-by: J. Bruce Fields <bfields@redhat.com>

If we end up doing "goto out_nomem" in this function, we'll call
nfsd_reply_cache_shutdown. That will attempt to walk the LRU list and
free entries, but that list may not be initialized yet if the server is
starting up for the first time. It's also possible for the shrinker to
kick in before we've initialized the LRU list.

Rearrange the initialization so that the LRU list_head and cache size
are initialized before doing any of the allocations that might fail.

Signed-off-by: Jeff Layton <jlayton@redhat.com>
Signed-off-by: J. Bruce Fields <bfields@redhat.com>
Upstream commit: ac534ff2d5508bdff1358a55d88053da729ff46b

Signed-off-by: Jarod Wilson <jarod@redhat.com>