nfsd: eliminate one of the DRC cache searches
Patchwork-id: 58778
O-Subject: [RHEL6.5 PATCH 28/34] [BZ#681606] nfsd: eliminate one of the DRC cache searches
Bugzilla: 681606
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>
RH-Acked-by: J. Bruce Fields <bfields@redhat.com>

The most common case is to do a search of the cache, followed by an
insert. In the case where we have to allocate an entry off the slab,
then we end up having to redo the search, which is wasteful.

Better optimize the code for the common case by eliminating the initial
search of the cache and always preallocating an entry. In the case of a
cache hit, we'll end up just freeing that entry but that's preferable to
an extra search.

Signed-off-by: Jeff Layton <jlayton@redhat.com>
Signed-off-by: J. Bruce Fields <bfields@redhat.com>
Upstream commit: 0b9ea37f24e247ed69baabf27fb211aa6a3e7622

Signed-off-by: Jarod Wilson <jarod@redhat.com>