nfsd: keep stats on worst hash balancing seen so far
Patchwork-id: 58782
O-Subject: [RHEL6.5 PATCH 32/34] [BZ#681606] nfsd: keep stats on worst hash balancing seen so far
Bugzilla: 681606
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>
RH-Acked-by: J. Bruce Fields <bfields@redhat.com>

The typical case with the DRC is a cache miss, so if we keep track of
the max number of entries that we've ever walked over in a search, then
we should have a reasonable estimate of the longest hash chain that
we've ever seen.

With that, we'll also keep track of the total size of the cache when we
see the longest chain. In the case of a tie, we prefer to track the
smallest total cache size in order to properly gauge the worst-case
ratio of max vs. avg chain length.

Signed-off-by: Jeff Layton <jlayton@redhat.com>
Signed-off-by: J. Bruce Fields <bfields@redhat.com>
Upstream commit: 98d821bda189ba2bfcb3877ea3064da3403698ae

Signed-off-by: Jarod Wilson <jarod@redhat.com>