ext4: fix online resizing for ext3-compat file systems
Patchwork-id: 58795
O-Subject: [PATCH RHEL6.5 1/2] ext4: fix online resizing for ext3-compat file systems
Bugzilla: 958939
RH-Acked-by: Lachlan McIlroy <lmcilroy@redhat.com>

For:

Bug 958939 - RHEL6.3: kernel panic during fsync, RIP ext4_mb_good_group,
after online resize of LUN on SAN from 500 to 800 GB

Backport of this upstream commit.  The commitlog doesn't
mention it, but it caused the problem in the above bug because
after resizing, we had an allocation goal block beyond the
un-extended "ngroups" limit, and it bypassed a loop control
mechanism, so we walked off the end of an array.

Tested with a yet-to-be-submitted xfstests testcase.

commit c5c72d814cf0f650010337c73638b25e6d14d2d4
Author: Theodore Ts'o <tytso@mit.edu>
Date:   Sun Apr 21 20:19:43 2013 -0400

    ext4: fix online resizing for ext3-compat file systems

    Commit fb0a387dcdc restricts block allocations for indirect-mapped
    files to block groups less than s_blockfile_groups.  However, the
    online resizing code wasn't setting s_blockfile_groups, so the newly
    added block groups were not available for non-extent mapped files.

    Reported-by: Eric Sandeen <sandeen@redhat.com>
    Signed-off-by: "Theodore Ts'o" <tytso@mit.edu>
    Cc: stable@vger.kernel.org

Signed-off-by: Jarod Wilson <jarod@redhat.com>